requirements.txt
python-telegram-bot
telethon
python-dotenv
honcho